requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
aiosqlite
httpx
nats-py
